military base factory

+++ src/main/kotlin/eu/hexgate/alternativeworld/domain/militarybase/EnergyGenerationData.kt
@@ -2,10 +2,10 @@ package eu.hexgate.alternativeworld.domain.militarybase
 
 data class EnergyGenerationData(
         val value: Int,
-        val rate: Int
+        val rate: Float
 ) {
 
-    fun nextLevel() = EnergyGenerationData(value * rate, rate)
+    fun nextLevel() = EnergyGenerationData((value * rate).toInt(), rate)
 
     fun calculateGeneration(weatherRate: Float) = (weatherRate * value).toInt()
